requirements.txt
fastapi
uvicorn
pymongo
mangum
python-dotenv
passlib[bcrypt]
pyjwt